# api/docs/swagger.yaml
definitions:
...
  structs.Categories:
    properties:
      categories:
        items:
          $ref: '#/definitions/structs.Category'
        type: array
    type: object
  structs.Category:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  structs.CategoryCreateReq:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  structs.Drive_Unit:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  structs.DriveUnitCreateReq:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  structs.DriveUnits:
    properties:
      drive_units:
        items:
          $ref: '#/definitions/structs.Drive_Unit'
        type: array
    type: object
  structs.Error:
    properties:
      message:
        type: string
    type: object
...
  structs.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/structs.Error'
    type: object
...
  structs.UpdateCategory:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
...
  structs.UpdateDriveUnit:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
...
info:
  contact: {}
paths:
...
  /v1/categories:
    get:
      consumes:
      - application/json
      description: This API for getting all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Categories'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: GetAllCategory
      tags:
      - category
  /v1/category:
    post:
      consumes:
      - application/json
      description: This API for creating category
      parameters:
      - description: CreateCategoryReq
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/structs.CategoryCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: CreateCategory
      tags:
      - category
    put:
      consumes:
      - application/json
      description: This API for updating category
      parameters:
      - description: UpdateCategory
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/structs.UpdateCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: UpdateCategory
      tags:
      - category
  /v1/category/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting category by id
      parameters:
      - description: Category_Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: DeleteCategory
      tags:
      - category
    get:
      consumes:
      - application/json
      description: This API for getting category by id
      parameters:
      - description: Category_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: GetCategory
      tags:
      - category
...
  /v1/drive:
    post:
      consumes:
      - application/json
      description: This API for creating drive unit
      parameters:
      - description: Drive_unit
        in: body
        name: drive
        required: true
        schema:
          $ref: '#/definitions/structs.DriveUnitCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Drive_Unit'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: CreateDiriveUnit
      tags:
      - drive_unit
    put:
      consumes:
      - application/json
      description: This API for updating drive_unit
      parameters:
      - description: UpdateDriveUnit
        in: body
        name: drive
        required: true
        schema:
          $ref: '#/definitions/structs.UpdateDriveUnit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Drive_Unit'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: UpdateDriveUnit
      tags:
      - drive_unit
  /v1/drive/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting drive_unit by id
      parameters:
      - description: Drive_Unit_Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Drive_Unit'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: DeleteDriveUnit
      tags:
      - drive_unit
    get:
      consumes:
      - application/json
      description: This API for getting drive_unit by id
      parameters:
      - description: Drive_Unit_Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Drive_Unit'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: GetDriveUnit
      tags:
      - drive_unit
  /v1/drives:
    get:
      consumes:
      - application/json
      description: This API for getting all drive_units
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.DriveUnits'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: GetAllDriveUnit
      tags:
      - drive_unit
...
swagger: "2.0"